refactor : initial setting 수정
