+++ src/main/java/ssumc/stardust/src/controller/UserController.java
@@ -1,8 +1,9 @@
-package ssumc.stardust.src.user;
+package ssumc.stardust.src.controller;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import ssumc.stardust.src.service.UserService;
 import ssumc.stardust.utils.JwtService;
 
 @RestController
