+++ src/main/java/ssumc/stardust/src/domain/Member.java
@@ -1,0 +1,15 @@
+package ssumc.stardust.src.domain;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
+
+@Entity
+public class Member {
+
+    @Id
+    @GeneratedValue
+    @Column(name = "member_id")
+    private Long id;
+}
