+++ src/main/java/ssumc/stardust/src/repository/UserRepository.java
@@ -1,0 +1,13 @@
+package ssumc.stardust.src.repository;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Repository;
+
+import javax.persistence.EntityManager;
+
+@Repository
+@RequiredArgsConstructor
+public class UserRepository {
+
+    private final EntityManager em;
+}
