--- src/main/java/ssumc/stardust/src/user/UserRepository.java
@@ -1,17 +1,0 @@
-package ssumc.stardust.src.user;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.stereotype.Repository;
-
-import javax.sql.DataSource;
-
-@Repository
-public class UserRepository {
-    private JdbcTemplate jdbcTemplate;
-
-    @Autowired
-    public void setDataSource(DataSource dataSource) {
-        this.jdbcTemplate = new JdbcTemplate(dataSource);
-    }
-}
