--- src/main/java/ssumc/stardust/src/user/model/PostSignUpReq.java
@@ -1,9 +1,0 @@
-package ssumc.stardust.src.user.model;
-
-import lombok.Data;
-
-@Data
-public class PostSignUpReq {
-    private String studentId;
-    private String phoneNum;
-}
